+++ lib/view/list_page.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class ListPage extends StatefulWidget {
+  const ListPage({super.key, required this.title});
+  final String title;
+
+  @override
+  State<ListPage> createState() => _ListPage();
+}
+
+class _ListPage extends State<ListPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: Center(child: Column(mainAxisAlignment: MainAxisAlignment.center)),
+    );
+  }
+}
